requirements.txt
asgiref==3.5.2
autopep8==1.6.0
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
psycopg2==2.9.5
pycodestyle==2.8.0
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
tzdata==2022.1
xlrd==2.0.1
xlwt==1.3.0
PS C:\Users\majag\wotc\wotc> cd..
PS C:\Users\majag\wotc> pip freeze
asgiref==3.5.2
autopep8==1.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.6
django-admin-interface==0.22.1
django-admin-list-filter-dropdown==1.0.3
django-betterforms==2.0.0
django-bootstrap-modal-forms==2.2.0
django-bootstrap5==22.2
django-ckeditor==6.5.1
django-colorfield==0.7.2
django-crispy-forms==1.14.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==3.0.1
django-jquery==3.1.0
django-js-asset==2.0.0
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
psycopg2==2.9.5
pycodestyle==2.8.0
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
tzdata==2022.1
xlrd==2.0.1
xlwt==1.3.0